requirements_test.txt
# Essential dependencies for testing
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing Libraries
pytest>=7.4.0
pytest-mock>=3.14.0
pytest-asyncio>=0.21.0

# Machine Learning
scikit-learn>=1.3.0

# Technical Analysis (if available)
# ta>=0.10.2

# Data processing
pyarrow>=12.0.0